refactor(auth): move auth modules into domain folders
