+++ client/src/lib/auth/authRedirect.js
@@ -1,7 +1,3 @@
-export function redirectToAuth(router, endpoint) {
-  router.push("/auth");
-}
-//uno solo para asduasbdu
-export function redirectToUnauthorized(router) {
-  router.push("/unauthorized");
+export function redirectTo(router, endpoint) {
+  router.push(endpoint);
 }
